pyproject.toml
[tool.poetry]
name = "NoodleBot"
version = "0.1.0"
description = "A small discord bot for pcparadise"
# all the people who've developed on this project can be found
# on the github repo, only the original author is written below.
authors = ["Arvinder Dhanoa <awsomearvinder@gmail.com>"]
maintainers = [ "Ganoodles" ]
# readme = "README.md"
license = "GPLv3"
packages = [
  { include = "src" }
]

[tool.poetry.dependencies]
python = "^3.10"
discord-py-interactions = "^5.5.0"
python-dotenv = "^1.0.0"
aiosqlite = "^0.19.0"
pathlib = "^1.0.1"

[tool.poetry.dev-dependencies]
black = "22.3.0"
pylint = "^2.14.5"
mypy = "~0.942"

[tool.poetry.scripts]
main = "src.main:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
